+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 - An owner field in the Todo schema which links a todo to a user.
 - A very basic template for Todo index.
 - A view for Todos.
+- A list todos API, (ie. Macroy.list_todos/1).
 
 ## [1.2.3] - 2019-11-28
 ### Fixed